backend/tools/acquire_paper/requirements.txt
boto3
requests
lxml
python-dotenv